+++ src/index_notation/index_notation_printer.cpp
@@ -66,11 +66,8 @@ void IndexNotationPrinter::visit(const LiteralNode* op) {
     case Datatype::Float64:
       os << op->getVal<double>();
       break;
-      case Datatype::Float128:
-          os << op->getVal<dd_real>();
-          break;
-      case Datatype::Float256:
-          os << op->getVal<qd_real>();
+      case Datatype::DDReal:
+          os << op->getVal<ddreal>();
           break;
     case Datatype::Complex64:
       os << op->getVal<std::complex<float>>();
@@ -78,11 +75,8 @@ void IndexNotationPrinter::visit(const LiteralNode* op) {
     case Datatype::Complex128:
       os << op->getVal<std::complex<double>>();
       break;
-      case Datatype::Complex256:
-          os << op->getVal<std::complex<dd_real>>();
-          break;
-      case Datatype::Complex512:
-          os << op->getVal<std::complex<qd_real>>();
+      case Datatype::DDComplex:
+          os << op->getVal<std::complex<ddreal>>();
           break;
     case Datatype::Undefined:
       break;