+++ src/ir/ir_printer.cpp
@@ -101,13 +101,9 @@ void IRPrinter::visit(const Literal* op) {
       stream << ((op->getValue<double>()!=0.0)
                  ? util::toString(op->getValue<double>()) : "0.0");
     break;
-      case Datatype::Float128:
-          stream << ((op->getValue<dd_real>()!=dd_real(0.0))
-                     ? util::toString(op->getValue<dd_real>()) : "0.0");
-          break;
-      case Datatype::Float256:
-          stream << ((op->getValue<dd_real>()!=dd_real(0.0))
-                     ? util::toString(op->getValue<dd_real>()) : "0.0");
+      case Datatype::DDReal:
+          stream << ((op->getValue<ddreal>()!=ddreal("0"))
+                     ? util::toString(op->getValue<ddreal>()) : "0");
           break;
     case Datatype::Complex64: {
       std::complex<float> val = op->getValue<std::complex<float>>();
@@ -119,13 +115,8 @@ void IRPrinter::visit(const Literal* op) {
       stream << val.real() << " + I*" << val.imag();
     }
     break;
-      case Datatype::Complex256: {
-          std::complex<dd_real> val = op->getValue<std::complex<dd_real>>();
-          stream << val.real() << " + I*" << val.imag();
-      }
-          break;
-      case Datatype::Complex512: {
-          std::complex<qd_real> val = op->getValue<std::complex<qd_real>>();
+      case Datatype::DDComplex: {
+          std::complex<ddreal> val = op->getValue<std::complex<ddreal>>();
           stream << val.real() << " + I*" << val.imag();
       }
           break;