+++ src/storage/typed_index.cpp
@@ -25,12 +25,10 @@ size_t TypedIndex::getAsIndex(const IndexTypeUnion& mem) const {
     case Datatype::Int128:
     case Datatype::Float32:
     case Datatype::Float64:
-    case Datatype::Float128:
-    case Datatype::Float256:
+    case Datatype::DDReal:
     case Datatype::Complex64:
     case Datatype::Complex128:
-    case Datatype::Complex256:
-    case Datatype::Complex512:
+    case Datatype::DDComplex:
     case Datatype::Undefined: taco_ierror; return 0;
   }
   taco_unreachable;
@@ -52,12 +50,10 @@ void TypedIndex::set(IndexTypeUnion& mem, const IndexTypeUnion& value) {
     case Datatype::Int128:
     case Datatype::Float32:
     case Datatype::Float64:
-    case Datatype::Float128:
-    case Datatype::Float256:
+    case Datatype::DDReal:
     case Datatype::Complex64:
     case Datatype::Complex128:
-    case Datatype::Complex256:
-    case Datatype::Complex512:
+    case Datatype::DDComplex:
     case Datatype::Undefined: taco_ierror;
   }
 }
@@ -77,12 +73,10 @@ void TypedIndex::setInt(IndexTypeUnion& mem, const int value) {
     case Datatype::Int128:
     case Datatype::Float32:
     case Datatype::Float64:
-    case Datatype::Float128:
-    case Datatype::Float256:
+    case Datatype::DDReal:
     case Datatype::Complex64:
     case Datatype::Complex128:
-    case Datatype::Complex256:
-    case Datatype::Complex512:
+    case Datatype::DDComplex:
     case Datatype::Undefined: taco_ierror;
   }
 }
@@ -102,12 +96,10 @@ void TypedIndex::add(IndexTypeUnion& result, const IndexTypeUnion& a, const Inde
     case Datatype::Int128:
     case Datatype::Float32:
     case Datatype::Float64:
-    case Datatype::Float128:
-    case Datatype::Float256:
+    case Datatype::DDReal:
     case Datatype::Complex64:
     case Datatype::Complex128:
-    case Datatype::Complex256:
-    case Datatype::Complex512:
+    case Datatype::DDComplex:
     case Datatype::Undefined: taco_ierror;
   }
 }
@@ -127,12 +119,10 @@ void TypedIndex::addInt(IndexTypeUnion& result, const IndexTypeUnion& a, const i
     case Datatype::Int128:
     case Datatype::Float32:
     case Datatype::Float64:
-    case Datatype::Float128:
-    case Datatype::Float256:
+    case Datatype::DDReal:
     case Datatype::Complex64:
     case Datatype::Complex128:
-    case Datatype::Complex256:
-    case Datatype::Complex512:
+    case Datatype::DDComplex:
     case Datatype::Undefined: taco_ierror;
   }
 }
@@ -152,12 +142,10 @@ void TypedIndex::multiply(IndexTypeUnion& result, const IndexTypeUnion& a, const
     case Datatype::Int128:
     case Datatype::Float32:
     case Datatype::Float64:
-    case Datatype::Float128:
-    case Datatype::Float256:
+    case Datatype::DDReal:
     case Datatype::Complex64:
     case Datatype::Complex128:
-    case Datatype::Complex256:
-    case Datatype::Complex512:
+    case Datatype::DDComplex:
     case Datatype::Undefined: taco_ierror;
   }
 }
@@ -177,12 +165,10 @@ void TypedIndex::multiplyInt(IndexTypeUnion& result, const IndexTypeUnion& a, co
     case Datatype::Int128:
     case Datatype::Float32:
     case Datatype::Float64:
-    case Datatype::Float128:
-    case Datatype::Float256:
+    case Datatype::DDReal:
     case Datatype::Complex64:
     case Datatype::Complex128:
-    case Datatype::Complex256:
-    case Datatype::Complex512:
+    case Datatype::DDComplex:
     case Datatype::Undefined: taco_ierror;
   }
 }
@@ -244,12 +230,10 @@ void TypedIndexVal::set(TypedComponentVal value) {
     case Datatype::Int128:
     case Datatype::Float32:
     case Datatype::Float64:
-    case Datatype::Float128:
-    case Datatype::Float256:
+    case Datatype::DDReal:
     case Datatype::Complex64:
     case Datatype::Complex128:
-    case Datatype::Complex256:
-    case Datatype::Complex512:
+    case Datatype::DDComplex:
     case Datatype::Undefined: taco_ierror; return;  }
 }
 
@@ -269,12 +253,10 @@ void TypedIndexVal::set(TypedComponentRef value) {
     case Datatype::Int128:
     case Datatype::Float32:
     case Datatype::Float64:
-    case Datatype::Float128:
-    case Datatype::Float256:
+    case Datatype::DDReal:
     case Datatype::Complex64:
     case Datatype::Complex128:
-    case Datatype::Complex256:
-    case Datatype::Complex512:
+    case Datatype::DDComplex:
     case Datatype::Undefined: taco_ierror; return;  }
 }
 
@@ -464,12 +446,10 @@ bool operator>(const TypedIndexVal& a, const TypedIndexVal &other) {
     case Datatype::Int128:
     case Datatype::Float32:
     case Datatype::Float64:
-    case Datatype::Float128:
-    case Datatype::Float256:
+    case Datatype::DDReal:
     case Datatype::Complex64:
     case Datatype::Complex128:
-    case Datatype::Complex256:
-    case Datatype::Complex512:
+    case Datatype::DDComplex:
     case Datatype::Undefined: taco_ierror; return false;
   }
   taco_unreachable;
@@ -492,12 +472,10 @@ bool operator==(const TypedIndexVal& a, const TypedIndexVal &other) {
     case Datatype::Int128:
     case Datatype::Float32:
     case Datatype::Float64:
-    case Datatype::Float128:
-    case Datatype::Float256:
+    case Datatype::DDReal:
     case Datatype::Complex64:
     case Datatype::Complex128:
-    case Datatype::Complex256:
-    case Datatype::Complex512:
+    case Datatype::DDComplex:
     case Datatype::Undefined: taco_ierror; return false;
   }
   taco_unreachable;
@@ -535,12 +513,10 @@ bool operator>(const TypedIndexVal& a, const int other) {
     case Datatype::Int128:
     case Datatype::Float32:
     case Datatype::Float64:
-    case Datatype::Float128:
-    case Datatype::Float256:
+    case Datatype::DDReal:
     case Datatype::Complex64:
     case Datatype::Complex128:
-    case Datatype::Complex256:
-    case Datatype::Complex512:
+    case Datatype::DDComplex:
     case Datatype::Undefined: taco_ierror; return false;
   }
   taco_unreachable;
@@ -562,12 +538,10 @@ bool operator==(const TypedIndexVal& a, const int other) {
     case Datatype::Int128:
     case Datatype::Float32:
     case Datatype::Float64:
-    case Datatype::Float128:
-    case Datatype::Float256:
+    case Datatype::DDReal:
     case Datatype::Complex64:
     case Datatype::Complex128:
-    case Datatype::Complex256:
-    case Datatype::Complex512:
+    case Datatype::DDComplex:
     case Datatype::Undefined: taco_ierror; return false;
   }
   taco_unreachable;